requirements.txt
Django==5.2.*
psycopg2-binary>=2.9.9
djangorestframework>=3.15.2
django-cors-headers>=4.4.0
python-dotenv>=1.0.1
whitenoise>=6.7.0
openpyxl>=3.1.5
reportlab>=4.2.2
pytest>=8.0.0
pytest-django>=4.8.0
pytest-cov>=5.0.0
ruff>=0.13.0
mkdocs-material
pygments
mkdocs
